Move character.ts to actors folder
