+++ module/actors/bwactor.ts
@@ -6,7 +6,7 @@ import { ArmorRootData } from "../items/armor.js";
 import { PossessionRootData } from "../items/possession.js";
 import { ReputationDataRoot } from "../items/reputation.js";
 import { TraitDataRoot, Trait } from "../items/trait.js";
-import { BWCharacterData } from "../character.js";
+import { BWCharacterData } from "./character.js";
 import { NpcData } from "../actors/npc.js";
 import { AffiliationDataRoot } from "../items/affiliation.js";
