+++ module/actors/character.ts
@@ -1,8 +1,8 @@
-import { Common, DisplayProps, ClumsyWeightData, TracksTests, BWActorDataRoot, Ability, BWActor } from "./actors/bwactor.js";
-import { CharacterBurnerDialog } from "./dialogs/characterBurner.js";
-import { ShadeString, TestString, canAdvance, updateTestsNeeded, getWorstShadeString, StringIndexedObject } from "./helpers.js";
-import { BWItemData } from "./items/item.js";
-import { Skill } from "./items/skill.js";
+import { Common, DisplayProps, ClumsyWeightData, TracksTests, BWActorDataRoot, Ability, BWActor } from "./bwactor.js";
+import { CharacterBurnerDialog } from "../dialogs/characterBurner.js";
+import { ShadeString, TestString, canAdvance, updateTestsNeeded, getWorstShadeString, StringIndexedObject } from "../helpers.js";
+import { BWItemData } from "../items/item.js";
+import { Skill } from "../items/skill.js";
 
 export class BWCharacter extends BWActor{
     data: CharacterDataRoot;
