+++ module/actors/sheets/character-sheet.ts
@@ -7,7 +7,7 @@ import {
 import { handleRollable } from "../../rolls/rolls.js";
 import { CharacterBurnerDialog } from "../../dialogs/characterBurner.js";
 import { addNewItem } from "../../dialogs/importItemDialog.js";
-import { BWCharacter } from "../../character.js";
+import { BWCharacter } from "../character.js";
 import { byName } from "../../helpers.js";
 import { ArmorRootData } from "../../items/armor.js";
 import { MeleeWeaponRootData, MeleeWeaponData } from "../../items/meleeWeapon.js";
