+++ module/rolls/rerollFate.ts
@@ -2,7 +2,7 @@ import { TestString } from "../helpers.js";
 import { BWActor, TracksTests } from "../actors/bwactor.js";
 import * as helpers from "../helpers.js";
 import { getNoDiceErrorDialog, RerollMessageData, rollDice, templates } from "./rolls.js";
-import { BWCharacter } from "../character.js";
+import { BWCharacter } from "../actors/character.js";
 import { Armor } from "../items/armor.js";
 import { Skill } from "../items/skill.js";
