+++ module/rolls/rollSkill.ts
@@ -16,7 +16,7 @@ import {
     RollOptions,
     mergeDialogData, getSplitPoolText, getSplitPoolRoll
 } from "./rolls.js";
-import { BWCharacter } from "../character.js";
+import { BWCharacter } from "../actors/character.js";
 import { Skill } from "../items/skill.js";
 import { Possession, PossessionRootData } from "../items/possession.js";
