+++ module/rolls/rollSpell.ts
@@ -1,5 +1,5 @@
 import { BWActor } from "../actors/bwactor.js";
-import { BWCharacter } from "../character.js";
+import { BWCharacter } from "../actors/character.js";
 import { Skill } from "../items/skill.js";
 import { Spell } from "../items/spell.js";
 import * as helpers from "../helpers.js";
