+++ module/rolls/rollWeapon.ts
@@ -3,7 +3,7 @@ import { handleSkillRoll } from "./rollSkill.js";
 import * as helpers from "../helpers.js";
 import { EventHandlerOptions, mergePartials, RollDialogData, RollOptions } from "./rolls.js";
 import { BWActor } from "../actors/bwactor.js";
-import { BWCharacter } from "../character.js";
+import { BWCharacter } from "../actors/character.js";
 import { RangedWeapon } from "../items/rangedWeapon.js";
 import { MeleeWeapon } from "../items/meleeWeapon.js";
 import { Skill } from "../items/skill.js";
